Move test utils to separate folder
