+++ src/ec-evaluator/__tests__/class-decl.test.ts
@@ -6,7 +6,7 @@ import {
   createContextStub,
   getControlItemStr,
   getStashItemStr
-} from "./utils";
+} from "./__utils__/utils";
 
 describe("evaluate NormalClassDeclaration correctly", () => {
   it("evaluate NormalClassDeclaration correctly", () => {
