+++ src/ec-evaluator/__tests__/mtd-overriding.test.ts
@@ -6,7 +6,7 @@ import {
   createContextStub,
   getControlItemStr,
   getStashItemStr
-} from "./utils";
+} from "./__utils__/utils";
 
 describe("evaluate method overriding correctly", () => {
   it("should invoke overriden instance method", () => {
